Rename control file
